trace2: log fsync stats in trace2 rather than wrapper
As a library boundary, it makes more sense for wrapper.c to not call any
trace2 code, but instead be able to provide statistics if called
upon. Therefore, moving the trace2 logging code to trace2.[ch] allows
wrapper.c to also not be dependent on trace2.h and repository.h.