unit test: add basic example
Although this commit doesn't showcase unit tests running against
anything in Git, locally, I have a smaller set of Git files compiling
into a library with unit tests running against it using this C TAP
harness.

However, you can run `make ctap` to get an idea of how the output looks
and use it as a baseline to play around with the other features of the
library.

Signed-off-by: Calvin Wan <calvinwan@google.com>
Change-Id: I8cc5c121dc67218692d6c11d0c0e7050a3de4bef